review_ai/pyproject.toml
[tool.poetry]
name = "review_ai"
version = "0.1.0"
description = ""
authors = ["mdfarhanali34 <farhandxb96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
pygithub = "^2.4.0"
flask = "^3.0.3"
gunicorn = "^23.0.0"
python-dateutil = "^2.9.0.post0"
openai = "^1.54.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
